Deploy stage optimization

+++ NetCoreSsh/SshDeployer.cs
@@ -1,22 +1,32 @@
-﻿using Renci.SshNet;
+﻿using System;
+using System.IO;
+using System.Threading.Tasks;
+using Renci.SshNet;
 using Serilog;
 
 namespace DotNetSsh
 {
     public class SshDeployer
     {
-        public void Deploy(string source, DeploymentOptions options)
+        public async Task Deploy(DirectoryInfo source, DeploymentOptions options, bool cleanTargetFolder)
         {
-            var sftp = new SftpClient(options.Host, options.Credentials.User, options.Credentials.Password);
-            var ssh = new SshClient(options.Host, options.Credentials.User, options.Credentials.Password);
-
-            using (var clients = new Clients(sftp, ssh))
+            try
             {
+                Log.Information("Starting deployment...");
+
+                var sftp = new SftpClient(options.Host, options.Credentials.User, options.Credentials.Password);
+                var ssh = new SshClient(options.Host, options.Credentials.User, options.Credentials.Password);
+
+                using var clients = new Clients(sftp, ssh);
                 clients.Connect();
 
-                SyncFiles(source, options, clients);
+                await SyncFiles(source, options, clients, cleanTargetFolder);
                 GiveExecutablePermission(options, clients);
                 RunIfSelected(options, clients.SshClient);
+            }
+            catch (Exception e)
+            {
+                throw new DeploymentException("Deployment failed", e);
             }
         }
 
@@ -46,19 +56,23 @@ namespace DotNetSsh
             return options.DestinationPath + "/" + options.AssemblyName;
         }
 
-        private static void SyncFiles(string source, DeploymentOptions options,
-            Clients clients)
+        private static Task SyncFiles(DirectoryInfo source, DeploymentOptions options,
+            Clients clients, bool cleanTargetFolder)
         {
             Log.Information("Deploying files...");
 
-            if (clients.SftpClient.Exists(options.DestinationPath))
+            PrepareTargetFolder(options, clients, cleanTargetFolder);
+            var synchronizer = new SshFolderSynchronizer(clients.SftpClient);
+            return synchronizer.Sync(source, options.DestinationPath);
+        }
+
+        private static void PrepareTargetFolder(DeploymentOptions options, Clients clients, bool cleanTargetFolder)
+        {
+            if (cleanTargetFolder && clients.SftpClient.Exists(options.DestinationPath))
             {
                 Log.Verbose("The destination folder already exists. We are going to delete it.");
                 clients.SshClient.DeleteExisting(options.DestinationPath);
             }
-
-            clients.SshClient.MirrorDirTree(options.DestinationPath, source);
-            clients.SftpClient.Copy(source, options.DestinationPath);
         }
     }
 }
